0.0.2: Initial release of the on-chain program
It provides five instructions to support the
queued multisig transfer.  The signers checks
all the proposed transfers and call approve
transactions.  Once the m/n threshold met,
all the transfers will be executed by the
program automatically.

Please refer to the Rust doc as well as the
TypeScript test for more detail.